switch from timestamp number to Date (wip)

+++ src/index.ts
@@ -49,10 +49,10 @@ export class Index {
    */
   id: string;
 
-  /** Hash map of hashes and files. Empty by default, and filled
+  /** Hash map of added files that were written to the odb. Empty by default, and filled
    * after [[Index.writeFiles]] has been called and the hashes of the files have been calculated.
    */
-  processed = new Map<string, FileInfo>();
+  processedAdded = new Map<string, FileInfo>();
 
   /**
    * A set of filepaths of new files that will be part of the new commit.
@@ -75,19 +75,21 @@ export class Index {
    * or can be useful to discard any added or deleted files from the index object.
    */
   invalidate(): Promise<void> {
-    // check if index exists, this can be false if the commit has no files (--allowEmpty)
-    return io.pathExists(this.getAbsPath()).then((exists: boolean) => {
-      if (exists) { return fse.unlink(this.getAbsPath()); }
-    }).then(() => {
-      this.repo.removeIndex(this);
+    if (this.repo) {
+      // check if index exists, this can be false if the commit has no files (--allowEmpty)
+      return io.pathExists(this.getAbsPath()).then((exists: boolean) => {
+        if (exists) { return fse.unlink(this.getAbsPath()); }
+      }).then(() => {
+        this.repo.removeIndex(this);
 
-      this.addRelPaths = new Set();
-      this.deleteRelPaths = new Set();
-      this.processed.clear();
-      this.id = undefined;
-      this.repo = null;
-      this.odb = null;
-    });
+        this.addRelPaths = new Set();
+        this.deleteRelPaths = new Set();
+        this.processedAdded.clear();
+        this.id = undefined;
+        this.repo = null;
+        this.odb = null;
+      });
+    }
   }
 
   /**
@@ -129,20 +131,22 @@ export class Index {
   private save(): Promise<void> {
     this.throwIfNotValid();
 
-    const userData: string = JSON.stringify({
-      adds: this.addRelPaths,
-      deletes: this.deleteRelPaths,
-      processed: this.processed,
-    }, (key, value) => {
-      if (value instanceof Map) {
-        return Array.from(value.entries());
-      }
-      if (value instanceof Set) {
-        return Array.from(value);
-      }
-      return value;
+    const processed = Array.from(this.processedAdded.entries()).map((res: [string, FileInfo]) => {
+      const size: number = res[1].stat.size;
+      const ctime: number = res[1].stat.ctime.getTime();
+      const mtime: number = res[1].stat.mtime.getTime();
+      return {
+        name: res[0], hash: res[1].hash, ext: res[1].ext, stat: { size, ctime, mtime },
+      };
     });
-    return fse.ensureDir(Index.getAbsDir(this.repo)).then(() => fss.writeSafeFile(this.getAbsPath(), userData))
+
+    const userData: string = JSON.stringify({
+      adds: Array.from(this.addRelPaths),
+      deletes: Array.from(this.deleteRelPaths),
+      processed,
+    });
+    return fse.ensureDir(Index.getAbsDir(this.repo))
+      .then(() => fss.writeSafeFile(this.getAbsPath(), userData))
       .then(() => this.repo.modified());
   }
 
@@ -168,9 +172,20 @@ export class Index {
         const index = new Index(repo, odb, isMainIndex ? '' : indexName.substr(6, indexName.length - 6)); // set 'abc123' as index id
         const content: string = parseIndex[1].toString();
         const json: any = JSON.parse(content);
+
         index.addRelPaths = new Set(json.adds);
         index.deleteRelPaths = new Set(json.deletes);
-        index.processed = new Map(json.processed);
+
+        index.processedAdded = new Map(json.processed.map((item: any) => {
+          const size: number = item.stat.size;
+          const ctime: Date = new Date(item.stat.ctime);
+          const mtime: Date = new Date(item.stat.mtime);
+          return [item.name, {
+            hash: item.hash,
+            ext: item.ext,
+            stat: { size, ctime, mtime },
+          }];
+        }));
         parseIndexes.push(index);
       }
       return parseIndexes;
@@ -190,7 +205,7 @@ export class Index {
 
       // if the file has already been processed from a previous 'index add .',
       // we don't need to do it again
-      if (!this.processed.has(relPath)) {
+      if (!this.processedAdded.has(relPath)) {
         this.addRelPaths.add(relPath);
       }
     }
@@ -207,7 +222,7 @@ export class Index {
     for (const filepath of filepaths) {
       const relPath: string = isAbsolute(filepath) ? relative(this.repo.workdir(), filepath) : filepath;
 
-      if (!this.processed.has(relPath)) {
+      if (!this.processedAdded.has(relPath)) {
         this.deleteRelPaths.add(relPath);
       }
     }
@@ -218,10 +233,10 @@ export class Index {
   /**
    * Hashes of files. Filled after [[Index.writeFiles]] has been called.
    */
-  getProcessedMap(): Map<string, FileInfo> {
+  getFileProcessedMap(): Map<string, FileInfo> {
     this.throwIfNotValid();
 
-    return this.processed;
+    return this.processedAdded;
   }
 
   /**
@@ -238,7 +253,7 @@ export class Index {
       .then(() => {
         const relPaths: string[] = difference(Array.from(this.addRelPaths), Array.from(this.deleteRelPaths));
 
-        unprocessedRelItems = relPaths.filter((p: string) => !this.processed.has(p));
+        unprocessedRelItems = relPaths.filter((p: string) => !this.processedAdded.has(p));
 
         if (flags & WRITE.SKIP_FILELOCK_CHECKS) {
           return Promise.resolve();
@@ -259,7 +274,7 @@ export class Index {
 
         // TODO: (Seb) Handle deleted files as well here
         for (const r of res.results) {
-          this.processed.set(r.file, r.fileinfo);
+          this.processedAdded.set(r.file, r.fileinfo);
         }
         return this.save();
       });
